clean code.
